Add files via upload
Unicorn Companies Data Visualization and Analysis -This project is aimed at analyzing and visualizing data of unicorn companies, which are private startups with a valuation of over $1 billion. The data is obtained from various sources, including Crunchbase, and includes information such as company name, valuation, funding, and industry.

Steps Taken Data Collection: The was provided for me by Quantum Analytics Ng for my first Python course project.
Data Cleaning: The next step was Manipulation and Validation of Data and transforming the data into a format for easy analysis.
Data Analysis: I calculated the Return On investment (ROI) for each company, sorted the companies by ROI, and identified the top 10 unicorn companies.
Data Visualization:  I created bar charts of the top 10 unicorn companies with their funding values , i also created line charts  with their Age of Joining for Each Company, i created pie charts of the 5 top countries that have the most unicorns and finally  I created column charts of the Top Investors in Funding Unicorns with their number of unicorns funded.
Challenges Faced One of the main challenges faced in this project was dealing with KEYERROR MESSAGE, like ['Age of Joining'] which i created in new column on the dataset when performing the cleaning of the data was not defined. To be able to know How long does it usually take for a company to become a unicorn? 
Overall, this project was successful in analyzing and visualizing data of unicorn companies.